+++ src/common/POFields/FieldIsManualPO.js
@@ -3,14 +3,21 @@ import PropTypes from 'prop-types';
 import { Field } from 'react-final-form';
 import { FormattedMessage } from 'react-intl';
 
-import { Checkbox } from '@folio/stripes/components';
+import { Checkbox, InfoPopover } from '@folio/stripes/components';
+
+const label = (
+  <>
+    <FormattedMessage id="ui-orders.orderDetails.manualPO" />
+    <InfoPopover content={<FormattedMessage id="ui-orders.orderDetails.manualPO.info" />} />
+  </>
+);
 
 const FieldIsManualPO = ({ disabled }) => {
   return (
     <Field
       component={Checkbox}
       fullWidth
-      label={<FormattedMessage id="ui-orders.orderDetails.manualPO" />}
+      label={label}
       name="manualPo"
       type="checkbox"
       disabled={disabled}